Refactoring - add*() functions no longer change state of buttons
The addACERelations(), addNERTags() and addPOSTags() functions no
longer change the state of buttons when they are called, since it
should be possible to add these visualizations to a page that does not
have any buttons.  The button state is now changed by the callback
functions attached to the buttons themselves.