BRAT visualizations now use token string length instead of textSpans
Labels for words in the BRAT sentence visualizations were not being
displayed properly because the textSpan string offsets were not being
set properly by the current version of the AGIGA2 pipeline.  The
changes in this commit circumvent the problem.

Sentences now displayed by the BRAT controls are now created by
joining the token.text field of each token using spaces, instead of
using the textSpan string offsets for the original document text in
comm.text.  String offsets for labels are now computed from the token
length, instead of using the token.textSpan offsets.  This approach
adds whitespace between every token, which makes the text fuglier.
While the previous version of Quicklime would have displayed this
sentence as is:

  John's daughter Mary expressed sorrow.

this version of Quicklime will display the sentence as:

  John 's daughter Mary expressed sorrow .