BRAT sentence <div>'s are now wrapped in a container <div>
Prep work for addressing issue #4, which requests labels for BRAT
sentences.

When BRAT adds an <svg> canvas to a <div>, it sets the height of that
<div> to the same number of pixels as the height of the <svg> canvas.
Instead of adding a label to the (fixed height) <div> that contains
the (fixed height) <svg> canvas, we wrap that (fixed height) <div> in
a container <div>.  The NER and POS buttons now show/hide the
container.

The DOM structure for displaying BRAT sentences was:

  <div class="brat_sentence" id="sentence_ner_[SENTENCE_UUID]">
  <div class="brat_sentence" id="sentence_pos_[SENTENCE_UUID]">

but is now:

  <div class="brat_sentence_container" id="sentence_ner_container_[SENTENCE_UUID]">
    <div class="brat_sentence" id="sentence_ner_[SENTENCE_UUID]">
  <div class="brat_sentence_container" id="sentence_pos_container_[SENTENCE_UUID]">
    <div class="brat_sentence" id="sentence_pos_[SENTENCE_UUID]">