DOM class 'entity_[ENTITY_UUID]' now added by QL.addCommunication().
The DOM class 'entity_[ENTITY_UUID]' was being added to token <span>'s
by the QL.addEntityTable() function, which is an optional function if
you just want to display the token text of a Communication without
adding the list of Entities and their EntityMentions.

Moved this functionality from QL.addEntityTable() to
QL.addCommunication().