TokenTaggings and Parses now associated with Tokenizations instead of Sentences
The DOM elements for TokenTaggings and Parses were associated (through
class names) with Sentence UUID's - but in the Concrete schema, the
TokenTagging, DependencyParse and Parse data structures are tied to
Tokenizations.  Identifying the DOM elements for these datastructures
by Sentence UUID is ambiguous when a Sentence has more than one
Tokenization.