Now using jQuery SVG files from a fork of the project
We were using a customized version of jQuery SVG that had some
features disabled for compatibility with newer versions of jQuery.  We
are now using the jQuery SVG files from a fork of the project:

  https://github.com/twitwi/svg

which actually fixes the compatibility issues with newer versions of
jQuery instead of just disabling unused functionality.