Improvements to how quicklime.brat handles TokenTaggings.
QL.addNERTags() and QL.addPOSTags() now assign BRAT "entity IDs" based
on the TaggedToken.tokenIndex variable, and not the location of that
TaggedToken in the TokenTagging's TaggedTokenList.  This prevents
alignment issues from occurring if TaggedTokens are added or removed
from the list of TaggedTokens.

For clarity, renamed nerTagList to nerTokenTagging, and posTagList to
posTokenTagging.