Properly handle new default value of -1 for Dependency.gov
The root node in a dependency parse was previously indicated by a null
Dependency.gov field:

  1: optional i32 gov        // will be null for ROOT token (only)

but the Concrete specification was recently changed so that
Dependency.gov has a default value of -1:

  1: optional i32 gov = -1  // can be omitted when dep is the root token

For Communications created by concrete-agiga, the root node in a
dependency parse now has a Dependency.gov value of -1.

Updated the visualization code to correctly handle Dependency.gov
values of -1, to prevent problems with Dagre errors about indexing
into an array with -1.