Sentences without (optional) textSpan field are now displayed properly
If a Concrete Sentence object is missing the optional
sentence.textSpan field, the character offsets for the sentence will
now be computed (if possible) from the character offsets of the tokens
in that sentence's tokenization.

Fixes an issue where the BRAT visualization would not be displayed
properly for this category of sentence object.